ruby: Merge jtdowney/aes_siv into the Ruby implementation
