+++ ruby/lib/sivchain.rb
@@ -4,7 +4,8 @@ require "openssl"
 
 require "sivchain/version"
 
-require "sivchain/cmac"
+require "sivchain/aes/siv"
+require "sivchain/aes/cmac"
 
 # Advanced symmetric encryption using the AES-SIV (RFC 5297) and CHAIN constructions
 module SIVChain
